requirements.txt
boto3
python-decouple
uuid
Flask-API
flask-restful
